+++ aeneas/tools/execute_job.py
@@ -8,6 +8,7 @@ as a container + configuration string (wizard case).
 
 import sys
 
+import aeneas.globalfunctions as gf
 from aeneas.executejob import ExecuteJob
 from aeneas.logger import Logger
 from aeneas.tools import get_rel_path
@@ -19,7 +20,7 @@ __copyright__ = """
     Copyright 2015,      Alberto Pettarin (www.albertopettarin.it)
     """
 __license__ = "GNU AGPL 3"
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 __email__ = "aeneas@readbeyond.it"
 __status__ = "Production"
 
@@ -54,6 +55,11 @@ def main():
 
     logger = Logger(tee=verbose)
     executor = ExecuteJob(logger=logger)
+
+    if not gf.can_run_c_extension():
+        print "[WARN] Unable to load Python C Extensions"
+        print "[WARN] Running the slower pure Python code"
+        print "[WARN] See the README file for directions to compile the Python C Extensions"
 
     print "[INFO] Loading job from container..."
     result = executor.load_job_from_container(container_path, config_string)